Can now pass the initial tasklist in the URL; added direct links
You can now go directly to a certain tasklist by adding /tasklist/{id}
to the URL. This sets a 'start' class on the given tasklist, and then
Javascript handles scrolling directly to this list when the page loads.

Also fixed some URLs in the layout template, and removed whitespace
between tasklists.