Finally removed the hardcoded callback url
This was annoying because I was testing this on a couple of different
servers and I'd have to change the callback uri each time I switched.

Moved the authentication check outside out of the index() method and
into a before_request method. This way callback url can be set using
url_for before the GTasks object is initialized. This means the GTasks
object is reinstantiated on every request, but I don't think that
increases the number of API calls, so that's ok.